Moved games logic from App.js, and component to the Components folder
